+++ lib/src/data/tilt_data.dart
@@ -144,20 +144,53 @@ class TiltDataModel {
       );
 }
 
-/// 倾斜 Stream
+/// 倾斜 Stream 数据
 @immutable
 class TiltStream {
   const TiltStream({
     required this.position,
-    required this.gesturesType,
-    this.enableRevert,
+    this.gesturesType = GesturesType.controller,
+    this.gestureUse = true,
   });
 
-  /// 当前坐标
+  /// 当前触发的坐标位置
+  ///
+  /// 会触发对应位置的倾斜效果
+  ///
+  /// 比如：
+  /// 组件尺寸 width: 10, height: 10，
+  /// - [position] 为 (0, 0) 坐标时，会触发最左上的倾斜
+  /// - [position] 为 (10, 10) 坐标时，会触发最右下的倾斜
   final Offset position;
 
+  /// 触发手势类型
+  ///
+  /// 手势优先级：
+  /// [GesturesType.touch] > [GesturesType.hover] > [GesturesType.controller] > [GesturesType.sensors]
+  ///
+  /// 如果需要使用动画或其他方式自行控制，
+  ///
+  /// *推荐* 使用 [GesturesType.controller]
+  ///
+  /// 如果使用其他的类型进行触发，
+  /// 那么就会受到对应类型相关配置、效果的影响。
+  ///
+  /// 比如：
+  /// 自行触发 [GesturesType.sensors] 的时候，
+  /// 配置 [TiltConfig.enableSensorRevert] 为 false 的情况下，
+  /// 将同样不会复原至初始状态。
   final GesturesType gesturesType;
 
-  /// 是否开启复原
-  final bool? enableRevert;
+  /// 手势是否正在使用
+  ///
+  /// 用于确定手势是否正在使用，
+  /// 并根据手势优先级进行处理。
+  ///
+  /// 比如：
+  /// 如果在触发 [GesturesType.touch] 的时候永远不赋值为 false，
+  /// 那么优先级低于 [GesturesType.touch] 的手势将永远不会被触发。
+  ///
+  /// - [true]  手势正在使用
+  /// - [false] 手势离开或不再使用
+  final bool gestureUse;
 }